Add demo Ideas DS
